API Reference

PhysicalBounds.build_COSMO_model_dual

PhysicalBounds.build_COSMO_model_primal

PhysicalBounds.build_Hypatia_model_dual

PhysicalBounds.build_Hypatia_model_primal

PhysicalBounds.construct_matrices

PhysicalBounds.dual_problem_solve

PhysicalBounds.heuristic_solve

PhysicalBounds.optimize

PhysicalBounds.optimize!

PhysicalBounds.primal_problem_solve

PhysicalBounds.to_complex

PhysicalBounds.to_real

PhysicalBounds.unvec_symm

PhysicalBounds.unvec_symm

PhysicalBounds.unvec_symm_scs

PhysicalBounds.vec_symm

PhysicalBounds.vec_symm_scs

PhysicalBounds.build_COSMO_model_dual — Method

build_COSMO_model_dual(
    C::AbstractMatrix{T}, 
    A::Vector{S}, 
    b::AbstractVector{T};
    x0=nothing, 
    y0=nothing,
    tol=1e-4, 
    verbose=false
) where {T, S <: Union{AbstractMatrix, AbstractSparseMatrix}}

Given an SDP in LMI form

\[\begin{aligned} & \text{minimize} && b^Ty \\ & \text{subject to} && \sum_{i=1}^m y_iA_i + C \succeq 0 \\ \end{aligned}\]

builds and returns a COSMO model.

The model is warmstarted at primal-dual point (x0, y0) and set to have stopping tolerance tol.

PhysicalBounds.build_COSMO_model_primal — Method

build_COSMO_model_primal(
    C::AbstractMatrix{T}, 
    A::Vector{S}, 
    b::AbstractVector{T};
    ineq=false,
    x0=nothing,
    y0=nothing,
    tol=1e-4,
    verbose=false
) where {T, S <: Union{AbstractSparseMatrix, AbstractMatrix}}

Given an SDP in standard form

\[\begin{aligned} & \text{maximize} && \mathbf{tr}(CX) \\ & \text{subject to} && \mathbf{tr}(A_iX) = b_i, \qquad i = 1, \dots, m \\ &&& X \succeq 0, \end{aligned}\]

builds and returns a COSMO model. If ineq is true, then the equality constraints are transformed into inequality constraints $\mathbf{tr}(A_iX) \le b$.

The model is warmstarted at primal-dual point (x0, y0) and set to have stopping tolerance tol.

PhysicalBounds.build_Hypatia_model_dual — Method

build_Hypatia_model_dual(
    C::AbstractMatrix{T},
    A::Vector{S},
    b::AbstractVector{T}
) where {T <: Real, S <: AbstractMatrix{T}}

Given an SDP in LMI form

\[\begin{aligned} & \text{minimize} && b^Ty \\ & \text{subject to} && \sum_{i=1}^m y_iA_i + C \succeq 0 \\ \end{aligned}\]

builds and returns a Hypatia model.

PhysicalBounds.build_Hypatia_model_primal — Method

build_Hypatia_model_primal(
    C::AbstractMatrix{T},
    A::Vector{S},
    b::AbstractVector{T}
) where {T <: Real, S <: Union{AbstractSparseMatrix, AbstractMatrix}}

Given an SDP in standard form

\[\begin{aligned} & \text{maximize} && \mathbf{tr}(CX) \\ & \text{subject to} && \mathbf{tr}(A_iX) = b_i, \qquad i = 1, \dots, m \\ &&& X \succeq 0, \end{aligned}\]

builds and returns a Hypatia model.

PhysicalBounds.construct_matrices — Method

contruct_matrices(physics::IntegralEquation, P::M, Q::M) where {M <: AbstractMatrix}

Constructs the matrices $\bar P, \bar Q, \bar A_i$ as in [Angeris et al. 2021] used to build the bounding semidefinite program.

PhysicalBounds.dual_problem_solve — Method

dual_problem_solve(C, Av, b; optimizer=Hypatia.Optimizer(verbose=false))

Solves the 'dual' SDP

\[\begin{aligned} & \text{minimize} && b^Ty \\ & \text{subject to} && \sum_{i=1}^m y_iA_i + C \succeq 0 \\ \end{aligned}\]

where the $A_i$'s are stored in the vector Av, using the specified optimizer.

Returns the model and variable $y$, after running the solver.

PhysicalBounds.heuristic_solve — Method

function heuristic_solve(
    ::Type{LBFGS}, 
    physics::IntegralEquation, 
    c_complex;
    θ₀=nothing,
    λ=0,
    rseed=0,
    m=20,
    factr=1e7,
    pgtol=1e-8,
    iprint=1,
    maxfun=1000,
    maxiter=250
)

Uses L-BFGS-B to solve the physical design problem

\[\begin{aligned} & \text{minimize} &&-f(z) \\ & \text{subject to} && A(\theta)z = b \\ &&& 0 \le \theta \le 1. \end{aligned}\]

The equality constraint is eliminated by substituting $A(\theta)^{-1}b$ for $z$. The objective $f$ is

\[f(z) = \frac{\lvert c^Tz\rvert}{\|z\|^2} + \lambda\lvert c^Tz\rvert.\]

Note that if we set $\lambda = 0$, we are optimizing for 'purity', whereas if we set $\lambda = \infty$, we are optimizing for power (in a particular mode).

The solver warmstarts at θ₀. The L-BFGS-B solver has parameters m, factr, pgtol, iprint, maxfun, and maxiter. See code for details.

PhysicalBounds.optimize! — Method

optimize!(model::COSMO.Workspace) = COSMO.optimize!(model)

Calls the COSMO solver on the problem model.

PhysicalBounds.optimize — Method

optimize(model::Hypatia.Models.Model{T}; verbose=true) where {T}

Calls the Hypatia solver on the problem model.

PhysicalBounds.primal_problem_solve — Method

primal_problem_solve(C, Av, b; ineq = nothing, optimizer=Hypatia.Optimizer(verbose=false), warmstart_X=nothing)

Solves the SDP

\[\begin{aligned} & \text{minimize} && \mathbf{tr}(CX) \\ & \text{subject to} && \mathbf{tr}(A_iX) = b_i, \qquad i = 1, \dots, m \\ &&& X \succeq 0, \end{aligned}\]

where the $A_i$'s are stored in the vector Av, using the specified optimizer.

Returns the model and variable $X$, after running the solver.

PhysicalBounds.to_complex — Function

to_complex(x)

Returns a complex vector y corresponding to x, where x is assumed to have the form

\[ x = \begin{bmatrix} \mathbf{Re}(y) \\ \mathbf{Im}(y) \end{bmatrix}.\]

PhysicalBounds.to_real — Function

to_real(x)

Converts x to real.

If x is a Vector, returns the real and imaginary parts stacked:[real(x)ᵀ imag(x)ᵀ]ᵀ`.

If x is a matrix, returns [real(x) -imag(x) ; imag(x) real(x)]

PhysicalBounds.unvec_symm — Method

unvec_symm(x, dim)

Returns a dim-by-dim symmetric matrix corresponding to x. x is a vector of length dim*(dim + 1)/2, corresponding to a symmetric matrix

X = [ X11  X12 ... X1k
      X21  X22 ... X2k
      ...
      Xk1  Xk2 ... Xkk ],

where vec(X) = (X11, X12, X22, X13, X23, ..., Xkk)

PhysicalBounds.unvec_symm — Method

unvec_symm(x)

Returns a dim-by-dim symmetric matrix corresponding to vector x, where dim is the unique positive solution to dim * (dim + 1) ÷ 2 = length(x):

X = [ X11  X12 ... X1k
      X21  X22 ... X2k
      ...
      Xk1  Xk2 ... Xkk ],

where vec(X) = (X11, X12, X22, X13, X23, ..., Xkk)

PhysicalBounds.unvec_symm_scs — Method

unvec_symm_scs(x)

Returns a dim-by-dim symmetric matrix corresponding to x. x is a vector of length dim*(dim + 1)/2, corresponding to a symmetric matrix

X = [ X11     X12/√2 ... X1k/√2
      X21/√2  X22    ... X2k/√2
      ...
      Xk1/√2  Xk2/√2 ... Xkk ],

where vec(X) = (X11, X12, X22, X13, X23, ..., Xkk)

Note that the factor √2 preserves inner products: x'*c = Tr(unvec_symm(c, dim) * unvec_symm(x, dim))

PhysicalBounds.vec_symm — Method

vec_symm(X)

Returns a vectorized representation of a symmetric matrix X. vec(X) = (X11, X12, X22, X13, X23, ..., Xkk)

PhysicalBounds.vec_symm_scs — Method

vec_symm_scs(X)

Returns a vectorized representation of a symmetric matrix X. vec(X) = (X11, √2*X12, X22, √2*X13, X23, ..., Xkk)

Note that the factor √2 preserves inner products: x'*c = Tr(unvec_symm(c, dim) * unvec_symm(x, dim))